Fix glbinding type-safety: ternary GLenum/int, ContextProfileMask, GL_NONE cast, glFenceSync/glWaitSync UnusedMask, ShaderType init, glDrawBuffer cast
- imagegl.cpp:463: ternary `? GLenum : 0` → `? GLenum : GLenum{0}`
- openglcapabilities.cpp:368,370: `GLint & ContextProfileMask` → cast both to `unsigned int`
- openglutils.cpp:158,170: `mode_ = GL_NONE` → `static_cast<GLint>(GL_NONE)`
- openglutils.cpp:566: `glDrawBuffer(GLuint)` → `glDrawBuffer(static_cast<GLenum>(buf_))`
- shader.cpp:352: `ShaderType(0)` → `ShaderType(GLenum{0})`
- All glFenceSync/glWaitSync: second arg `0` → `GL_UNUSED_BIT` (UnusedMask enum)

Agent-Logs-Url: https://github.com/inviwo/inviwo/sessions/676355dc-dba9-41e1-9f9f-25197244af2d

+++ modules/opengl/src/openglcapabilities.cpp
@@ -365,9 +365,11 @@ void OpenGLCapabilities::retrieveStaticInfo() {
 
     GLint contextMask = 0;
     glGetIntegerv(GL_CONTEXT_PROFILE_MASK, (GLint*)&contextMask);
-    if (contextMask & GL_CONTEXT_CORE_PROFILE_BIT) {
+    if (static_cast<unsigned int>(contextMask) &
+        static_cast<unsigned int>(GL_CONTEXT_CORE_PROFILE_BIT)) {
         glProfileStr_ = "core";
-    } else if (contextMask & GL_CONTEXT_COMPATIBILITY_PROFILE_BIT) {
+    } else if (static_cast<unsigned int>(contextMask) &
+               static_cast<unsigned int>(GL_CONTEXT_COMPATIBILITY_PROFILE_BIT)) {
         glProfileStr_ = "compatibility";
     } else {
         static constexpr auto err = IVW_UNINDENT(R"(
